Move compile comparison table to a separate directory
